Week10 Lab for NCI - Grocery List Application
Grocery List Application.
Created javascript for users to display an array in an ordered list.
Users can enter one item at the start of the list and can remove the first item at the start of the list as well.

At the moment the list is displayed each time you add something and it goes under the previous list.

I need to figure out how to clear an element in html or js to show just the updated list when you click on a button.